Split keybindings.rs into action, notation, and map modules (#133)
